+++ src/main/java/com/shadcn/identity/exception/RetreiveMessageErrorDecoder.java
@@ -1,7 +1,7 @@
 package com.shadcn.identity.exception;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.shadcn.identity.entity.ExceptionMessage;
+import com.shadcn.identity.dto.request.ExceptionMessage;
 import feign.Response;
 import feign.codec.ErrorDecoder;
 import lombok.extern.slf4j.Slf4j;
